requirements/laser.txt
absl-py==2.0.0
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Brotli
certifi
cffi
chardet==5.2.0
charset-normalizer
click==8.1.7
colorama==0.4.6
cryptography
DataProperty==1.0.1
datasets==2.14.6
dill==0.3.7
distro==1.8.0
einops==0.7.0
evaluate==0.4.1
fast-hadamard-transform==1.0.1
filelock
flash-attn==2.3.4
frozenlist==1.4.0
fsspec==2023.10.0
glog==0.3.1
gmpy2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
idna 
Jinja2
joblib==1.3.2
jsonlines==4.0.0
lm-eval==0.3.0
MarkupSafe
mbstrdecoder==1.1.3
mkl-fft 
mkl-random 
mkl-service==2.4.0
mpmath 
multidict==6.0.4
multiprocess==0.70.15
networkx 
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.8.7
numpy 
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.3.2
packaging==23.2
pandas==2.1.3
pathvalidate==3.2.0
peft==0.7.1
Pillow 
portalocker==2.8.2
primefac==2.0.12
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pybind11==2.11.1
pycountry==22.3.5
pycparser
pydantic==2.5.1
pydantic_core==2.14.3
pyOpenSSL 
PySocks 
pytablewriter==1.2.0
python-dateutil==2.8.2
python-gflags==3.1.2
pytz==2023.3.post1
PyYAML 
# quiptools-cuda==0.0.0
regex==2023.10.3
requests 
responses==0.18.0
rouge-score==0.1.2
sacrebleu==1.5.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sqlitedict==2.1.0
sympy
tabledata==1.3.3
tcolorpy==0.1.4
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
tqdm-multiprocess==0.0.11
transformers
# transformers==4.34.0
triton==2.1.0
typepy==1.3.2
typing_extensions
tzdata==2023.3
urllib3
xxhash==3.4.1
yarl==1.9.2
zstandard==0.22.0
